Fix/Enhancement: Permit lua files to be located away from application
Mudlet will now seek out and check for LuaGlobal.lua (which loads all
other Mudlet specific lua files) on start up and permit the user to search
for it if not found or is not read accessible (or is too large - see
below).  Before the search is invoked the user is advised what the problem
is and what to do and directions given as to what will happen if they both
can and cannot complete the required action.  Once a suitable instance of
LuaGlobal.lua is found it's location is stored in the Mudlet.conf file
and used in the future.  As of this commit if a wrong version is selected
the only way to change it is to hand delete the relevant line from that
human readable file.  The in-game console text shown during lua start-up
now shows the directory from which the Mudlet lua files were loaded and
then the working directory being used afterwards.

During Lua subsystem start up the lua search path is now set to include the
directory of LuaGlobal.lua and the current directory is temporarily changed
to that during loading of the Mudlet lua and Geyser lua files before being
changed to that of the Mudlet executable.  LuaGlobal.lua has been edited
to NOT now search for the wanted files it loads in some hard-coded
locations - it expects to be IN the required direction on invocation.

Adds four mudlet themed icons for general usage with QMessageBox and other
dialogs.

Deletes obsolete "all-in-one" LuaGlobal.lua file in root of src tree
which shares a name with the wanted one in a sub-directory.  The
check mentioned above only accepts a file of this name less than 10K Bytes
as the obsolete one is 60K and the wanted one is around 5K.

The need to provide a sensible default location to search for the
LuaGlobal.lua file for different operating systems had a useful side-effect
in that it also specifies the location for an installer to put those
files.  All the lua files required to operate Mudlet have been explicitly
brought into the Windows/Unix src.pro and *nix installer commands added
so that Mudlet can now be "deployed" after building in Qt Creator with
a "make install" executed in the build directory.

*** With these changes Qt "Shadow building" (out of source directory) works ***

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>