Enhancement: Provide limited *nix Install capability
With the preceding changes to the lua subsystem it is now possible to
attempt to provide a *nix make install capability within the Qt Creator
IDE.  However there are issues with the resulting Makefile which does
not comprehend that the user compiling the code may not have sufficient
privileges as is to actually install the component parts of Mudlet in a
(possibly) multi-user system.  This addition uses GNU awk (other flavours
might work) to carefully modify the qmake generated Makefile so that the
parts that add or remove files and directories from the main system area
(typically off of the /usr/local area of the filesystem) are run with
the necessary privilege escalation.  To do this it uses sudo and a utility
such as ssh-askpass that provides a GUI dialog for sudo to ask for a
password as required in a GUI environment such as the Qt IDE.

The target destinations are set by the .path elements of the target, and
LUA install targets set in the src.pro project file.  To maintain
consistency with values compiled into Mudlet the variable that should be
modified by the UNIX user in the project file is the fairly conventional,
according to Linux File Hierarchical Standard, PREFIX one that is set to
/usr/local near the top of the project file.


Signed-off-by: Stephen Lyons <slysven@virginmedia.com>