Revision: Removed unnecessary external unix install script
Found a simpler means to run "make install" when QtCreator user is not
normally privileged enough to do so (see src.pro project file for
instructions).

Also checked behaviour of qmake "install sets" when the .files member
is given just a path as a file to copy FROM - the entire tree from the
root of qmake's idea of the project's source directory (the Mudlet ./src
one) to each of the (effectively wild-carded) file in that path is
pre-pended to the .path member which is not what is wanted - SO THE FILES
MUST BE INDIVIDUALLY LISTED - which perhaps is a good thing in that odd
extra files in the source tree won't get included by accident.

Specifying files explicitly like this should still work with the mac
install process - so an intended merger of the mac_src.pro file into
the main project file should be possible.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>