Enhancement: Debug controls framework for developer run-time modifications
Provide places to put in code elements that place controls into the last
tab in the dlgProfilePreferences.cpp dialog that can be used to modify,
possibly temporarily, the operation of the program on a per host basis or
application wide.  Useful to allow the run-time selection of alternate
pieces of code for performance comparisons or to investigate suspect code.

The primary reason to inserting this framework is so that developers can
override the per-user setting of the Mudlet Lua files location - setting
the added check-box option on the last tab of the dlgProfilePreference
dialog forces following instances of Mudlet and any profiles opened after
in the current instance to use the Mudlet lua files in the src directory of
the Mudlet sources rather that the value that was set and remains stored in the user's
Mudlet configuration file.  Resetting the option reverts to normal
operation for following cases.  The setting IS retained and carries over
to following instances of the application.

Also included here is a spin-box control to reduce the alpha of the colour
used to draw the rooms on the 2D map, which I had done to help with
checking the placement of exits lines close to/under the room symbol.

The extra lines in the comments are intentional to allow sufficient space
between individual controls so that patching to insert or remove them
can be done automagically (the patch program has an easier time of it
if lines around changes remain stable through changes, or so I understand).

N.B. Host.cpp gains the Q_OBJECT macro so qmake WILL need to be run to
ensure it gets run though the "moc" when the make is done!

If the Todo plugin is available and enabled in QtCreator IDE one can locate
the places where the code is by adding the string "DEBUGCONTROLS" as a
title of one of things to track as this and a following spaced digit/letter
is embedded in the opening comment for each of the placeholders for
blocks of code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>