Update: Requested changes from other developers
Following a suggestion from another developer the code in T2DMap.cpp to
reduce the opacity of the room symbols has been removed from the 2D mapper
though the control variable code has been left - commented out - in place
as an example.

Following a request from Vadim, a developer override control has been added
to force the use of the Lua files distributed in the source package by
in-source or shadow-build instances of the Mudlet application - the setting
is stored between invocations of the application but any profiles loaded
(auto-login profiles or the first profile loaded) - before the profile
preferences dialog becomes available - will obviously use EITHER {if RESET}
the automatically maintained storage location for LuaGlobal.lua that Mudlet
rechecks each time it starts since the "fileLocation" modification was
incorporated OR {if SET} the source one DEPENDING ON the previously stored
value of this control - which is initially RESET in this patch.


Signed-off-by: Stephen Lyons <slysven@virginmedia.com>